% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors_Dev.R
\docType{data}
\name{lcrt_color_palettes}
\alias{lcrt_color_palettes}
\title{Full Palette List}
\format{
An object of class \code{list} of length 12.
}
\usage{
lcrt_color_palettes
}
\description{
A full list of the color palettes used in this package
}
\keyword{datasets}
