% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors_Dev.R
\docType{data}
\name{rpg_basemaps}
\alias{rpg_basemaps}
\title{lcrt Mapbox Base Maps}
\format{
An object of class \code{list} of length 3.
}
\usage{
rpg_basemaps
}
\description{
URLs to use Mapbox vector tile services as an interactive basemap
}
\keyword{datasets}
